Added find() method in HashTable

+++ cpp/lib_calvin/container/b_tree_base.h
@@ -148,7 +148,7 @@ public:
 	typedef T value_type;
 	typedef T * pointer;
 	typedef T & reference;
-	typedef int size_type;
+	typedef size_t size_type;
 	typedef ptrdiff_t difference_type;
 
 private:
@@ -160,7 +160,7 @@ private:
 		typedef T value_type;
 		typedef T const * pointer;
 		typedef T const & reference;
-		typedef int size_type;
+		typedef size_t size_type;
 		typedef ptrdiff_t difference_type;
 		typedef std::bidirectional_iterator_tag iterator_category;
 		reference operator*() const;
